Resolved linting issues
