+++ src/components/Toggle/index.ts
@@ -1,1 +1,3 @@
-export { Toggle } from './Toggle';
+import { Toggle } from './Toggle';
+
+export { Toggle };
